ActivitySelectionUnsortedUsingJava
implementing the logic for solving the problem of a activity selection for getting the maximum count of the activity can be done within the given time intervals by using the basics of Greedy algorithms and concepts.
Here we have sorted the data on the basis of end time of the activity given.