BinaryTree-uniValuedTreeCheck()
checking whether the given binary tree is having similar node value or not. If the binary tree nodes altogether are similar to each other then yes is to be printed otherwise, no is to be printed and each left node value from the root of the given binary tree and each right node value from the root of the given binary tree are compared with each other including the root node of that particular subtree of the given binary tree using java.