Heaps-minimumTimefillNslots()
program for getting the count of the instances that is required to visit each and every box individually in one go, whose indexes are provided in an array, with the total number of elements present in the array. This had been done by using LinkedList and Queue from java collection framework.